scripts: Clean up ganesha_mgr help
Use a multiline f-string to neaten the usage message.

Add a couple of functions to print help messages and exit.  Use them
to reduce duplication and increase consistency of both the content of
error messages and the code that produces them.

Consistently print the name of the actual script being run.

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Change-Id: I9da99cf52e37e496f45578c2a18cb1226a066ae2